// src/lib/schema.d.ts
export interface Database {
  public: {
    Tables: {
// ...
      users: {
        Row: {
          avatar_url: string | null
          bg_url: string | null
          birthdate: string | null
          cp_organizer_status: number | null
          created_at: string | null
          email: string | null
          id: string
          lastname: string | null
          name: string | null
          role: string | null
          updated_at: string | null
          username: string | null
        }
        Insert: {
          avatar_url?: string | null
          bg_url?: string | null
          birthdate?: string | null
          cp_organizer_status?: number | null
          created_at?: string | null
          email?: string | null
          id: string
          lastname?: string | null
          name?: string | null
          role?: string | null
          updated_at?: string | null
          username?: string | null
        }
        Update: {
          avatar_url?: string | null
          bg_url?: string | null
          birthdate?: string | null
          cp_organizer_status?: number | null
          created_at?: string | null
          email?: string | null
          id?: string
          lastname?: string | null
          name?: string | null
          role?: string | null
          updated_at?: string | null
          username?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      set_claim: {
        Args: {
          uid: string
          claim: string
          value: Json
        }
        Returns: string
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}